allow copy as fallback solution when resolution is bad
